# Conflicts:
#	aequilibrae/project/database_specification/transit/tables/table_list.txt
#	aequilibrae/project/database_specification/transit/tables/transit_pattern_links.sql
#	aequilibrae/project/database_specification/transit/tables/transit_pattern_mapping.sql
#	aequilibrae/project/database_specification/transit/tables/transit_routes.sql
#	aequilibrae/project/project.py
#	aequilibrae/transit/functions/__init__.py
#	aequilibrae/transit/functions/create_raw.py
#	aequilibrae/transit/functions/data.py
#	aequilibrae/transit/functions/db_utils.py
#	aequilibrae/transit/functions/del_pattern.py
#	aequilibrae/transit/functions/get_srid.py
#	aequilibrae/transit/gtfs_loader.py
#	aequilibrae/transit/lib_gtfs.py
#	aequilibrae/transit/map_matching_graph.py
#	aequilibrae/transit/transit.py
#	aequilibrae/transit/transit_elements/agency.py
#	aequilibrae/transit/transit_elements/link.py
#	aequilibrae/transit/transit_elements/pattern.py
#	aequilibrae/transit/transit_elements/route.py
#	aequilibrae/transit/transit_elements/stop.py
#	aequilibrae/transit/transit_elements/trip.py

+++ aequilibrae/project/project.py
@@ -57,7 +57,6 @@ class Project:
         file_name = os.path.join(project_path, "project_database.sqlite")
         if not os.path.isfile(file_name):
             raise FileNotFoundError("Model does not exist. Check your path and try again")
-
         self.project_base_path = project_path
         self.path_to_file = file_name
         self.source = self.path_to_file
@@ -91,7 +90,7 @@ class Project:
         self.activate()
 
         self.__create_empty_network()
-        self.__create_empty_transit()
+        self.create_empty_transit()
         self.__load_objects()
         self.about.create()
         global_logger.info(f"Created project on {self.project_base_path}")
@@ -187,7 +186,11 @@ class Project:
         self.conn.commit()
         initialize_tables(self, "network")
 
-    def __create_empty_transit(self):
+    def create_empty_transit(self):
+        """Creates the public transport database"""
+        if os.path.exists(os.path.join(self.project_base_path, "public_transport.sqlite")):
+            raise FileExistsError("Public Transport database already exists.")
+
         shutil.copyfile(spatialite_database, os.path.join(self.project_base_path, "public_transport.sqlite"))
         self.transit = Transit(self)
         initialize_tables(self, "transit")
